+++ classes/helper/ProductHelper.php
@@ -57,65 +57,6 @@ class ProductHelper
 		return '<p>' . $strKeywords . '</p>';
 	}
 	
-	/**
-	 * @param $objFile
-	 * @param $arrExifData
-	 *
-	 * @return string serialized multidimensional array
-	 */
-	public static function importImageToIsotopeMediamanager($objFile, $arrExifData, $imgSize = null)
-	{
-		if (($strSrc = self::createResizedImage($objFile, $arrExifData, $imgSize)) === null) {
-			return null;
-		}
-		
-		$arrFile = [
-			'src'       => $strSrc,
-			'alt'       => '',
-			'link'      => '',
-			'desc'      => '',
-			'translate' => 'none'
-		];
-		
-		return serialize([$arrFile]);
-	}
-	
-	/**
-	 * Calculate new dimensions and create image
-	 *
-	 * @param $objFile
-	 * @param $arrExifData
-	 *
-	 * @return null|string
-	 */
-	
-	protected static function createResizedImage($objFile, $arrExifData, $imgSize = null)
-	{
-		list($target, $size, $name) = ProductHelper::prepareImageStore($objFile, $arrExifData, $imgSize);
-		
-		$newImage = \Image::get($objFile->path, $size['size'][0], $size['size'][1], $size['size'][2], $target);
-		
-		$objNewFile = \Dbafs::addResource(urldecode($newImage));
-		
-		
-		return $name;
-	}
-	
-	
-	protected function prepareImageStore($objFile, $arrExifData, $imgSize)
-	{
-		$name = ltrim($objFile->name, '_');
-		$path = $objFile->path;
-		
-		if (!empty($imgSize)) {
-			$name = str_replace('.' . $objFile->extension, '_' . $imgSize['size'][0] . '.' . $objFile->extension, $name);
-			$path = str_replace($objFile->name, $name, $objFile->path);
-		} else {
-			$imgSize = ['size' => [$arrExifData['width'], $arrExifData['height'], 'center-center']];
-		}
-		
-		return [$path, $imgSize, $name];
-	}
 	
 	/**
 	 * create download element for each set size of isotope product image
@@ -127,7 +68,7 @@ class ProductHelper
 	public static function createDownloadItem($id, $file, $size)
 	{
 		$name = str_replace('.' . $file->extension, ProductHelper::getReplacer($file, $size), ltrim($file->name, '_'));
-		$path = str_replace($file->name, $name, $file->path);
+		$path = $downloadPath = str_replace($file->name, $name, $file->path);
 		
 		if (!file_exists($path)) {
 			$downloadPath = \Image::get($file->path, $size['size'][0], $size['size'][1], $size['size'][2], $path);
@@ -146,6 +87,7 @@ class ProductHelper
 		$objDownload->title     = $size['name'];
 		$objDownload->singleSRC = $downloadFile->uuid;
 		$objDownload->published = 1;
+		
 		$objDownload->save();
 		
 	}
@@ -168,56 +110,6 @@ class ProductHelper
 	}
 	
 	/**
-	 * @param $arrExif
-	 * @param $objFile
-	 *
-	 * @return string
-	 */
-	public static function generateAliasFromTitleOrFilename($arrExif, $objFile)
-	{
-		if ($arrExif['title']) {
-			$strAlias = $arrExif['title'];
-		} else {
-			$strAlias = str_replace('.' . $objFile->extension, '', $objFile->name);
-		}
-		
-		return General::generateAlias('', $objFile->id, 'tl_iso_product', $strAlias);
-	}
-	
-	/**
-	 * @param $src string
-	 *
-	 * @return int|string    filesize in readable form
-	 */
-	public function getFileSize($src)
-	{
-		$bytes = filesize($src);
-		
-		if ($bytes >= 1073741824) {
-			$bytes = number_format($bytes / 1073741824, 2) . ' GB';
-		} elseif ($bytes >= 1048576) {
-			$bytes = number_format($bytes / 1048576, 2) . ' MB';
-		} elseif ($bytes >= 1024) {
-			$bytes = number_format($bytes / 1024, 2) . ' KB';
-		} elseif ($bytes > 1) {
-			$bytes = $bytes . ' bytes';
-		} elseif ($bytes == 1) {
-			$bytes = $bytes . ' byte';
-		} else {
-			$bytes = '0 bytes';
-		}
-		
-		return $bytes;
-	}
-	
-	public static function getFileTitle($uuid)
-	{
-		$file = Download::findBy('singleSRC', $uuid);
-		
-		return $file->title;
-	}
-	
-	/**
 	 * return all product groups that are defined as editable in module
 	 *
 	 * @param $module object
@@ -226,14 +118,17 @@ class ProductHelper
 	 */
 	public function getEditableCategories($module)
 	{
-		if ($module->iso_editableCategories) {
-			$categories = [];
-			
-			foreach (deserialize($module->iso_editableCategories) as $cat) {
-				$categories[$cat] = ProductType::findByPk($cat)->name;
-			}
-			
-			return $categories;
+		if(!$module->iso_editableCategories)
+			return [];
+		
+		$categories = [];
+
+		foreach (deserialize($module->iso_editableCategories) as $cat) {
+			$categories[$cat] = ProductType::findByPk($cat)->name;
 		}
+
+		asort($categories);
+		
+		return $categories;
 	}
 }